src/requirements.txt
certifi==2026.1.4
duckdb==1.4.3
fastparquet==2025.12.0
gcsfs==2025.12.0
google-cloud-storage==3.7.0
pandas==2.3.3
pyarrow==22.0.0
pytest==9.0.2
pytest-mock==3.15.1
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.5.0